videoplatform/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
contourpy==1.3.1
cycler==0.12.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.55.3
kiwisolver==1.4.7
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pygame==2.6.1
PyJWT==2.9.0
pyparsing==3.2.0
PySimpleGUI==5.0.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
rsa==4.9
scipy==1.14.1
six==1.17.0
sqlparse==0.5.3
